Add prune_bundler option
This option makes it possible to specify puma in a Gemfile, but then
have puma use future Gemfiles in future releases when those releases are
seen via cluster workers.

Basically, this implements the feature that many people want: the
ability to specify puma in a context file and have restarts pickup
Gemfile changes.