Fix bug setting worker_directory when using a symlink directory
The @cli_options variable is used to reflect the options passed from the command line.
The @options variable is the final set of options that will be used by the Runner instance
for configuration. When generating the restart data, we want to set the value in the @options
hash and not in the @cli_options hash.